updated datasets and odometry methods to accept a vector of IMUData. started working on switching from WNOA to WNOJ for steamLIO. create a class for creating bias interpolation, not done yet.

+++ steam_icp/include/steam_icp/imu.hpp
@@ -3,7 +3,7 @@
 #include <Eigen/Dense>
 
 struct IMUData {
-  uint64_t timestamp = 0;
+  double timestamp = 0;
   Eigen::Vector3d ang_vel = Eigen::Vector3d::Zero();
   Eigen::Vector3d lin_acc = Eigen::Vector3d::Zero();
 };
